use iterator to point to the current style
This avoids crash if one (or both) of groupMembershipsSequence and
documentPageShapeOrders come in empty (when reading broken file. I do
not know if it can happen legitimately. Probably not).

Also, the old code actually overwrites the first elements of the two
vectors on each change, because assignment to a reference changes the
object the reference refers to. While it is probably harmless, as we do
not need the old styles again, it does not "feel" right.