tdf#126292 Fix unit conversion
With this commit I have added unit support.

In .VSD format, all length values (points, kilometers, feets etc.) are stored in Inches,
all Elapsed time are stored in Days,
all angles are stored in radians.

It means that value should be converted from Inch/ Days or radians,
even if the unit is not displayed.

For example:
 - centimeter values should multipled by 2.54
 - elapsed hour values should be multiplied by 24
 - percent values should be multipled by 100

Change-Id: I88711a1bf1bfd6078dc6948194449bcc91f34cce
Reviewed-on: https://gerrit.libreoffice.org/75374
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>
Tested-by: Bartosz Kosiorek <gang65@poczta.onet.pl>