Update patch set 3
Patch Set 3:

Information from Fridrich, how to start implementation:
1. Create a connector in draw and export in flat odg and see what it does
try to design the api the way that the information needed will be in that propList
Then parse the information into it.
propList can be extended without breaking anything (librevenge).
Just in libodfgen, you will detect whether some property exists, and do connectors with it and if not, do normal shapes.

The API should be defined in VSDCollector.h header.

Patch-set: 3