Avoid undefined behaviour when the std::vector is empty
By getting address of first element (e.g. &tmpBuffer[0]),
in case vector is empty, the undefined behaviour could occur.

data() returns a pointer to the underlying array serving as element storage.
The pointer is such that range [data(), data() + size()) is always a valid range,
even if the container is empty (data() is not dereferenceable in that case).

Change-Id: I62665a11ece0cc7e2226608004c0eb48788f95d1
Reviewed-on: https://gerrit.libreoffice.org/c/libvisio/+/173740
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>