tdf#168475 libvisio: a bit better handling of vsdx files theme import
with using the correct VariationColorIndex and VariationStyleIndex
from the xml for selecting the correct color from theme.xml.

Using the default color (do not apply the theme color) is necessary
for the first shape of the grouped "End Event" object which is a
standard black color and cannot be changed in MS Visio. TODO: unit test

Change-Id: I6da23215eb89b4177d045155cfaa8969cc85af12
Reviewed-on: https://gerrit.libreoffice.org/c/libvisio/+/191099
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Miklos Vajna <vmiklos@collabora.com>